requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-phonenumber-field==8.1.0
django-resized==1.0.3
djangorestframework==3.16.1
drf-yasg==1.21.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
